+++ src/seed/tests/loader/c.js
@@ -1,5 +1,0 @@
-KISSY.add("c",function(){
-
-},{
-    requires:["./a"]
-});