+++ src/seed/tests/loader/cyclic/a.js
@@ -1,0 +1,5 @@
+KISSY.add("cyclic/a",function(){
+
+},{
+    requires:["./b"]
+});