+++ src/seed/tests/loader/cyclic/c.js
@@ -1,0 +1,5 @@
+KISSY.add("cyclic/c",function(){
+
+},{
+    requires:["./a"]
+});